requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
urllib3==2.2.2
numpy==2.0.1
opencv-python==4.10.0.84
pillow==10.4.0
pytesseract==0.3.10
matplotlib==3.9.1
